Update part_11__60_second_countdown.md
Might need to check the delay input code - but it definitely needs something here or the code just ends before you can see the you win/you lose message. Also changed 'loose' to 'lose'. It now doesn't match the sample code at the bottom but its close? Changes seem to be: compscore has changed to comp_score, loose to lose and the delay input message, which is commented out in the sample code.